restructure frontend
