+++ frontend/peerprep/components/addquestionform.tsx
@@ -19,12 +19,12 @@ import { WysiMarkEditor } from "./wysimarkeditor";
 import BoxIcon from "./boxicons";
 
 import { capitalize, languages } from "@/utils/utils";
-import { complexityColorMap } from "@/app/(default)/questions-management/columns";
+import { complexityColorMap } from "@/app/@application/questions-management/columns";
 import {
   useUniqueCategoriesFetcher,
   isValidQuestionSubmission,
   submitQuestion,
-} from "@/services/questionService";
+} from "@/app/api/services/questionService";
 
 interface AddQuestionFormProps {
   initialTitle?: string;
@@ -83,7 +83,7 @@ export default function AddQuestionForm({
   // Handle removing a category
   const removeCategory = (category: string) => {
     setCategories((prevCategories) =>
-      prevCategories.filter((cat) => cat !== category),
+      prevCategories.filter((cat) => cat !== category)
     );
   };
 
@@ -112,7 +112,7 @@ export default function AddQuestionForm({
   // Handle input change for test case
   const handleInputChange = (
     index: number,
-    event: React.ChangeEvent<HTMLInputElement>,
+    event: React.ChangeEvent<HTMLInputElement>
   ) => {
     const updatedTestCases = [...testCases];
 
@@ -154,11 +154,11 @@ export default function AddQuestionForm({
         description,
         category,
         templateCode,
-        testCases,
+        testCases
       )
     ) {
       setErrorMessage(
-        "Please fill in all the required fields before submitting.",
+        "Please fill in all the required fields before submitting."
       );
       setErrorModalOpen(true); // Show error modal with the validation message
 
@@ -175,7 +175,7 @@ export default function AddQuestionForm({
         category,
         selectedTab,
         templateCode,
-        testCases,
+        testCases
       );
 
       if (response.ok) {
@@ -187,14 +187,14 @@ export default function AddQuestionForm({
         const errorData = await response.json();
 
         setErrorMessage(
-          errorData.error || "Failed to submit the question. Please try again.",
+          errorData.error || "Failed to submit the question. Please try again."
         );
         setErrorModalOpen(true);
       }
     } catch (error) {
       // Show error modal with generic error message
       setErrorMessage(
-        "An error occurred while submitting the question. Please try again later",
+        "An error occurred while submitting the question. Please try again later"
       );
       setErrorModalOpen(true);
     }
