+++ frontend/peerprep/components/editquestionform.tsx
@@ -24,14 +24,14 @@ import { capitalize, languages } from "@/utils/utils";
 import {
   complexityColorMap,
   Question,
-} from "@/app/(default)/questions-management/columns";
+} from "@/app/@application/questions-management/columns";
 import {
   deleteQuestion,
   editQuestion,
   useQuestionDataFetcher,
   useUniqueCategoriesFetcher,
   isValidQuestionSubmission,
-} from "@/services/questionService";
+} from "@/app/api/services/questionService";
 
 interface EditQuestionFormProps {
   initialTitle?: string;
@@ -72,7 +72,7 @@ export default function EditQuestionForm({
   const [testCases, setTestCases] =
     useState<{ input: string; output: string }[]>(initialTestCases);
   const [questionToDelete, setQuestionToDelete] = useState<Question | null>(
-    null,
+    null
   );
   const [question, setQuestion] = useState<Question | null>(null);
 
@@ -96,8 +96,8 @@ export default function EditQuestionForm({
               .map((str) => str.trim());
 
             return { input, output };
-          }),
-        ) || [],
+          })
+        ) || []
       );
       setQuestion(questionData?.question);
     }
@@ -141,7 +141,7 @@ export default function EditQuestionForm({
   // Handle removing a category
   const removeCategory = (category: string) => {
     setCategories((prevCategories) =>
-      prevCategories.filter((cat) => cat !== category),
+      prevCategories.filter((cat) => cat !== category)
     );
   };
 
@@ -170,7 +170,7 @@ export default function EditQuestionForm({
   // Handle input change for test case
   const handleInputChange = (
     index: number,
-    event: React.ChangeEvent<HTMLInputElement>,
+    event: React.ChangeEvent<HTMLInputElement>
   ) => {
     const updatedTestCases = [...testCases];
     const { name, value } = event.target;
@@ -215,11 +215,11 @@ export default function EditQuestionForm({
         description,
         categories,
         templateCode,
-        testCases,
+        testCases
       )
     ) {
       setErrorMessage(
-        "Please fill in all the required fields before submitting.",
+        "Please fill in all the required fields before submitting."
       );
       setErrorModalOpen(true); // Show error modal with the validation message
 
@@ -234,7 +234,7 @@ export default function EditQuestionForm({
         categories,
         selectedTab,
         templateCode,
-        testCases,
+        testCases
       );
 
       if (response.ok) {
@@ -244,7 +244,7 @@ export default function EditQuestionForm({
         const errorData = await response.json();
 
         setErrorMessage(
-          errorData.error || "Failed to update the question. Please try again.",
+          errorData.error || "Failed to update the question. Please try again."
         );
         setErrorModalOpen(true);
       }
