+++ frontend/peerprep/components/questionstable.tsx
@@ -31,12 +31,12 @@ import {
   Question,
   RenderCell,
   complexityOptions,
-} from "@/app/(default)/questions-management/columns";
+} from "@/app/@application/questions-management/columns";
 import {
   deleteQuestion,
   useQuestionsFetcher,
   useUniqueCategoriesFetcher,
-} from "@/services/questionService";
+} from "@/app/api/services/questionService";
 
 export default function QuestionsTable() {
   const [errorMessage, setErrorMessage] = useState<string>("");
@@ -52,7 +52,7 @@ export default function QuestionsTable() {
     onOpenChange: onConfirmModalOpenChange,
   } = useDisclosure();
   const [questionToDelete, setQuestionToDelete] = useState<Question | null>(
-    null,
+    null
   );
 
   const handleDelete = (question: Question) => {
@@ -91,7 +91,7 @@ export default function QuestionsTable() {
     complexityFilter,
     categoryFilter,
     sortDescriptor,
-    page,
+    page
   );
 
   const { categoryData, categoryLoading } = useUniqueCategoriesFetcher();
