+++ frontend/peerprep/package.json
@@ -34,6 +34,7 @@
     "dotenv": "^16.4.5",
     "framer-motion": "~11.1.1",
     "intl-messageformat": "^10.5.0",
+    "iron-session": "^8.0.3",
     "next": "14.2.4",
     "next-remove-imports": "^1.0.12",
     "next-runtime-env": "^3.2.2",
